third_party/WebKit/LayoutTests/fast/canvas/canvas-font-ex-units-crash.html

<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<body>
<script>
test(function(t) {
    ctx = document.createElement('canvas').getContext('2d');
    
    ctx.font = "5ex sans-serif";
    
    size = parseInt(ctx.font.substr(0, 2));
    family = ctx.font.substr(5);
    
    assert_approx_equals(size, 25, 10);
    assert_equals(family, 'sans-serif');
}, "Test that setting a font with size in 'ex' units doesn't crash.");
</script>
</body>
